Mostrar alerta de confirmación

+++ app/Http/Controllers/Admin/FamilyController.php
@@ -87,7 +87,24 @@ class FamilyController extends Controller
      */
     public function destroy(Family $family)
     {
+        if ($family->categories->count() > 0) {
+
+            session()->flash('swal', [
+                'icon' => 'error',
+                'title' => '¡Ups!',
+                'text' => 'No se puede eliminar la familia porque tiene categorias asociadas'
+            ]);
+
+            return redirect()->route('admin.families.edit', $family);
+        }
+
         $family->delete();
+
+        session()->flash('swal', [
+                'icon' => 'success',
+                'title' => '¡Bien hecho!',
+                'text' => 'Familia eliminada correctamente.'
+            ]);
 
         return redirect()->route('admin.families.index');
     }
